refactor into new package and create new BatchReviewScanner
