+++ src/main/java/nwoolcan/model/brewery/production/batch/review/BatchReviewScanner.java
@@ -1,0 +1,4 @@
+package nwoolcan.model.brewery.production.batch.review;
+
+public interface BatchReviewScanner {
+}
