+++ src/main/java/nwoolcan/model/brewery/production/batch/review/types/AbstractBatchReviewType.java
@@ -1,0 +1,4 @@
+package nwoolcan.model.brewery.production.batch.review.types;
+
+public class AbstractBatchReviewType {
+}
